arreglados algunos bugs, borrar perfil

+++ src/app/catalogo/catalogo.component.ts
@@ -56,8 +56,12 @@ export class CatalogoComponent implements OnInit {
     //Debo encontrar la manera de enlazar la actiivdad de la lista elegida con esta otra pantalla
   }
   eliminarActivity(actividad: Actividad):void{
-    this.frontendService.deleteActividad(actividad).subscribe(act => this.ngOnInit(), err => console.error('Ops: ' + err.message));
+    var r = confirm("¿Estás seguro de que quieres borrar la actividad?");
+      if (r == true) {
+        this.frontendService.deleteActividad(actividad).subscribe(act => this.ngOnInit(), err => console.error('Ops: ' + err.message));
     
+    } else {
+    } 
   }
 
 goBack(): void {
